更新

+++ crm/src/main/java/com/linxi/mapper/PayrecordMapper.java
@@ -17,10 +17,10 @@ public interface PayrecordMapper {
     void savePayrecord(Payrecord payrecord);
 
     //根据成交客户编号查询支付总金额
-    Integer queryPPaysumBySId(Integer sId);
+    Double queryPPaysumBySId(Integer sId);
 
     //根据成交客户编号查询退款总金额
-    Integer queryRefundBySId(Integer sId);
+    Double queryRefundBySId(Integer sId);
 
     //根据筛选条件查询支付记录
     List<Payrecord> queryPByScreen(@Param("page") Integer page,
